refactor(server): #125 — extract direct-messaging handlers into routes/direct.rs
